Move ls tests to a separate module
